refactored authentication into a concern module.
