+++ app/controllers/concerns/authentication_concern.rb
@@ -1,0 +1,16 @@
+module AuthenticationConcern
+  extend ActiveSupport::Concern
+
+  included do
+    before_action :authenticate_user!
+  end
+
+  def authenticate_user!
+    raise Pundit::NotAuthorizedError, "You need to log in to access this page." unless user_signed_in?
+  end
+
+  def user_not_authorized
+    flash[:alert] = "You are not authorized to perform this action."
+    redirect_back fallback_location: root_url
+  end
+end
