+++ app/controllers/exams_controller.rb
@@ -1,6 +1,7 @@
 class ExamsController < ApplicationController
   before_action :set_exam, only: %i[ show edit update destroy ]
-  before_action :authenticate_user!
+  include AuthenticationConcern
+
   def index
     @exams = Exam.all
     authorize @exams
@@ -92,13 +93,5 @@ class ExamsController < ApplicationController
 
   def exam_params
     params.require(:exam).permit(:name, :max_num_questions, :max_duration)
-  end
-  def authenticate_user!
-    raise Pundit::NotAuthorizedError, "You need to log in to access this page." unless user_signed_in?
-  end
-
-  def user_not_authorized
-    flash[:alert] = "You are not authorized to perform this action."
-    redirect_back fallback_location: root_url
   end
 end
